+++ lib/contents/rtl_directional_languages.dart
@@ -1,0 +1,14 @@
+final kRtlTextDirectionalLanguages = [
+  "ar", // Arabic
+  "arc", // Aramaic
+  "dv", // Divehi (Dhivehi, Maldivian)
+  "fa", // Persian (Farsi)
+  "ha", // Hausa (with Arabic script)
+  "he", // Hebrew
+  "khw", // Khowar (with Arabic script)
+  "ks", // Kashmiri (with Arabic script)
+  "ku", // Kurdish (Sorani)
+  "ps", // Pashto
+  "ur", // Urdu
+  "yi" // Yiddish
+];