+++ lib/models.dart
@@ -1,6 +1,7 @@
 library;
 
+export 'models/country.dart';
+export 'models/language.dart';
 export 'models/remaining_duration.dart';
 export 'models/selection.dart';
 export 'models/selector.dart';
-export 'models/country.dart';