+++ lib/utils/text_parser.dart
@@ -112,11 +112,3 @@ class SpannedText extends SpanText {
     return '$SpannedText(text: $text, types: $types)';
   }
 }
-
-void main() {
-  String normalText =
-      "My name is Mr. X. I'm <CUSTOM_TAG_1><CUSTOM_TAG_2>24</CUSTOM_TAG_2></CUSTOM_TAG_1> old.";
-  List<SpanText> parsedText = TextParser.parse(normalText);
-  print(
-      parsedText); // OUTPUT: [NormalText(text: My name is Mr. X. I'm ), SpannedText(text: 24, types: [CUSTOM_TAG_1, CUSTOM_TAG_2]), NormalText(text:  old.)]
-}